Figured out the bsearch problem.
If we bsearch and don't find an exact match but do end the main loop in
the middle of the file, it means there is a line in the file we should
be at and it may not be the one we're currently on.

For example, if we're doing a SeekType::LastLessThan and there are lines
both greater than and less than the prefix, we need to leave the cursor
at the start of the last line that is less than the prefix. But when we
exit the main loop the current line may be the line after the line we
should be on.

Not fixed in this commit, just wanted to write this down.